feat: Move back to saucelabs
BREAKING CHANGE: Move back to sauce labs and generisize options to be platform agnostic. Closes #384.
