+++ package.json
@@ -16,11 +16,11 @@
     "cz-conventional-changelog": "1.2.0",
     "eslint": "3.5.0",
     "eslint-config-airbnb": "11.1.0",
+    "eslint-friendly-formatter": "^2.0.6",
     "eslint-plugin-import": "1.15.0",
     "eslint-plugin-jsx-a11y": "2.2.2",
     "eslint-plugin-mocha": "4.5.1",
     "eslint-plugin-react": "6.2.2",
-    "eslint-friendly-formatter": "^2.0.6",
     "install": "0.8.1",
     "karma": "1.3.0",
     "karma-chai-plugins": "0.8.0",
@@ -29,7 +29,8 @@
     "karma-mocha": "1.1.1",
     "karma-opera-launcher": "1.0.0",
     "karma-safari-launcher": "1.0.0",
-    "karma-testingbot-launcher": "0.1.2",
+    "karma-sauce-launcher": "^1.0.0",
+    "karma-saucelabs-launcher": "0.0.0",
     "karma-sourcemap-loader": "0.3.7",
     "karma-webpack": "1.8.0",
     "less": "2.7.1",
